Partly refactor generators submodule
