+++ dwave/graphs/generators/common/common.py
@@ -1,3 +1,19 @@
+# Copyright 2026 D-Wave
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+__all__ = ["_add_compatible_edges", "_add_compatible_nodes", "_add_compatible_terms"]
+
 
 def _add_compatible_edges(G, edge_list):
     # Check edge_list defines a subgraph of G and create subgraph.
@@ -11,6 +27,7 @@ def _add_compatible_edges(G, edge_list):
         if G.number_of_edges() < len(edge_list):
             raise ValueError('edge_list contains duplicates.')
 
+
 def _add_compatible_nodes(G, node_list):
     if node_list is not None:
         if not all(G.has_node(n) for n in node_list):
@@ -20,7 +37,8 @@ def _add_compatible_nodes(G, node_list):
         G.remove_nodes_from(remove_nodes)
         if G.number_of_nodes() < len(node_list):
             raise ValueError('node_list contains duplicates.')
-        
+
+
 def _add_compatible_terms(G, node_list, edge_list):
     _add_compatible_edges(G, edge_list)
     _add_compatible_nodes(G, node_list)
